wip: content arranger based on prompts
